This PR adds webauthn support
Chromium is phasing out U2F support in Feb 2022
https://developer.chrome.com/blog/deps-rems-95/

We need to move to the newer, better Webauthn standard.

Since the majority of our 2factor auths are stored in U2F format, we
will use a compatibility layer to verify. See:

https://www.stackallocated.com/blog/2019/u2f-to-webauthn/
https://github.com/cedarcode/webauthn-ruby/blob/master/docs/u2f_migration.md
https://github.com/hansnielsen/webauthn-rails-demo/blob/v1-sign-with-webauthn/app/controllers/two_factor_controller.rb

Note that this PR covers just signing in using the Webauthn to U2F
compatibility layer (by passing app_id in extensions). We still need to
register with Webauthn. That will require DB migrations and is the subject
of a second PR.

+++ app/controllers/u2f_authentications_controller.rb
@@ -5,27 +5,47 @@ class U2fAuthenticationsController < ApplicationController
   include TwoFactorAuth
 
   def create
-    u2f_response = U2F::SignResponse.load_from_json(params[:u2f_response])
     publisher = pending_2fa_current_publisher
+    client_json_wrapper = JSON.parse(params[:webauthn_u2f_response])
+    client_json_response = client_json_wrapper["response"]
+    domain = request.base_url
+    response_id = client_json_wrapper["id"]
 
-    registration = publisher.u2f_registrations.find_by(key_handle: u2f_response.key_handle)
+    registration = publisher.u2f_registrations.find_by(key_handle: response_id)
+
+    # For future: To determine if this is a U2F request:
+    # client_json_wrapper[:clientExtensionResults][:appid] is set if is
+
+    assertion_response = WebAuthn::AuthenticatorAssertionResponse.new(
+      user_handle: registration.key_handle,
+      authenticator_data: Base64.urlsafe_decode64(client_json_response["authenticatorData"]),
+      client_data_json: Base64.urlsafe_decode64(client_json_response["clientDataJSON"]),
+      signature: Base64.urlsafe_decode64(client_json_response["signature"])
+    )
 
     begin
-      u2f.authenticate!(
-        session[:challenge],
-        u2f_response,
-        Base64.decode64(registration.public_key),
-        registration.counter
+      assertion_response.verify(
+        Base64.urlsafe_decode64(session[:current_authentication][:challenge]),
+        domain,
+        public_key: Base64.urlsafe_decode64(registration.public_key),
+        sign_count: registration.counter,
+        rp_id: domain
       )
-    rescue U2F::Error => e
-      Rails.logger.debug("U2F::Error! #{e}")
+    rescue WebAuthn::VerificationError => e
+      Rails.logger.debug("WebAuthn::Error! #{e}")
       redirect_to two_factor_authentications_path
       return
     ensure
-      session.delete(:challenge)
+      session.delete(:current_authentication)
     end
 
-    registration.update!(counter: u2f_response.counter)
+    counter = assertion_response.authenticator_data.sign_count
+    if registration.counter >= counter
+      Rails.logger.debug("Credential replay detected!")
+      redirect_to two_factor_authentications_path
+      return
+    end
+    registration.update!(counter: counter)
     session.delete(:pending_2fa_current_publisher_id)
 
     sign_in(:publisher, publisher)
